fix: format update

+++ Coding_Problems/CPP/16_longestValidParanthesis.cpp
@@ -11,7 +11,7 @@ int findIsValidParanthesis(string s)
     st.push(-1);
     int maxLength = 0;
 
-    for (int i = 0; i < (int)s.length(); i++)
+    for (int i = 0; i < s.length(); i++)
     {
         if (s[i] == '(')
             st.push(i);
